refactored things
